server/dev-requirements.txt
Flask
libsass
requests
python-dotenv
psycopg2-binary
flask-jwt-extended